29-03-2018 00:40
Cambios de tipos de Datos, algoritmo Insertar Proveedor

+++ MELB_WS/Models/Inventario/Modelos/Proveedor.cs
@@ -12,8 +12,8 @@ namespace MELB_WS.Models.Inventario.Modelos
         public int ID_Proveedor { get; set; }
         [StringLength(15, MinimumLength = 1)]
         public string Nombre { get; set; }
-        public int Telefono_1 { get; set; }
-        public int Telefono_2 { get; set; }
+        public Decimal Telefono_1 { get; set; }
+        public Decimal Telefono_2 { get; set; }
         [StringLength(50, MinimumLength = 1)]
         public string Correo { get; set; }
         [StringLength(15, MinimumLength = 1)]
